Move encoding logic to its own module
